+++ chunjun-connectors/chunjun-connector-influxdb/src/main/java/com/dtstack/chunjun/connector/influxdb/converter/InfluxdbColumnConverter.java
@@ -30,9 +30,12 @@ import com.dtstack.chunjun.converter.IDeserializationConverter;
 import com.dtstack.chunjun.converter.ISerializationConverter;
 import com.dtstack.chunjun.element.AbstractBaseColumn;
 import com.dtstack.chunjun.element.ColumnRowData;
-import com.dtstack.chunjun.element.column.BigDecimalColumn;
 import com.dtstack.chunjun.element.column.BooleanColumn;
 import com.dtstack.chunjun.element.column.BytesColumn;
+import com.dtstack.chunjun.element.column.DoubleColumn;
+import com.dtstack.chunjun.element.column.FloatColumn;
+import com.dtstack.chunjun.element.column.IntColumn;
+import com.dtstack.chunjun.element.column.LongColumn;
 import com.dtstack.chunjun.element.column.NullColumn;
 import com.dtstack.chunjun.element.column.StringColumn;
 import com.dtstack.chunjun.element.column.TimestampColumn;
@@ -191,25 +194,25 @@ public class InfluxdbColumnConverter
             case INTEGER:
                 return val -> {
                     if ("JSON".equals(format)) {
-                        return new BigDecimalColumn(((Double) val).intValue());
+                        return new IntColumn(((Double) val).intValue());
                     }
-                    return new BigDecimalColumn((Integer) val);
+                    return new IntColumn((Integer) val);
                 };
             case FLOAT:
-                return val -> new BigDecimalColumn(((Double) val).floatValue());
+                return val -> new FloatColumn(((Double) val).floatValue());
             case DOUBLE:
-                return val -> new BigDecimalColumn((Double) val);
+                return val -> new DoubleColumn((Double) val);
             case VARBINARY:
                 return val -> new BytesColumn((byte[]) val);
             case BIGINT:
                 return val -> {
                     if ("JSON".equals(format)) {
-                        return new BigDecimalColumn(((Double) val).longValue());
+                        return new LongColumn(((Double) val).longValue());
                     }
                     if (val instanceof Long) {
-                        return new BigDecimalColumn((Long) val);
+                        return new LongColumn((Long) val);
                     } else {
-                        return new BigDecimalColumn((Integer) val);
+                        return new LongColumn((Integer) val);
                     }
                 };
